[docs]move Cash,Exchange files to screens
